requirements.txt
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
dateutils==0.6.6
decorator==4.2.1
Django==2.0.4
django-filter==2.0.0
djangorestframework==3.8.2
idna==2.6
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.3
pandas==0.22.0
parso==0.1.1
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
python-decouple==3.1
pytz==2018.4
requests==2.18.4
rows==0.3.1
simplegeneric==0.8.1
six==1.11.0
tqdm==4.22.0
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
